+++ src/world/Territory/Cell.cpp
@@ -25,8 +25,8 @@ Sapphire::Cell::~Cell()
 void Sapphire::Cell::init( uint32_t x, uint32_t y, TerritoryPtr pZone )
 {
   m_pZone = pZone;
-  m_posX = x;
-  m_posY = y;
+  m_posX = static_cast< uint16_t >( x );
+  m_posY = static_cast< uint16_t >( y );
 
   m_actors.clear();
 }
